pyproject.toml
[project]
name = "file-search-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.52.0",
    "python-dotenv>=1.0.0",
]
